src: slim down node_v8_platform-inl.h
